pyproject.toml
[project]
name = "medium-clone-suggestion"
version = "0.1.0"
description = ""
authors = [
    {name = "Eliot6001",email = "jilali2004@gmail.com"}
]
packages = [
  { include = "medium_clone_suggestion", from = "src" }
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "keybert (>=0.9.0,<0.10.0)",
    "transformers (>=4.50.3,<5.0.0)",
    "supabase (>=2.15.0,<3.0.0)",
    "sentence-transformers (>=4.0.1,<5.0.0)",
    "nltk (>=3.9.1,<4.0.0)",
    "psutil (>=7.0.0,<8.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "beautifulsoup4 (>=4.13.3,<5.0.0)",
    "gensim (>=4.3.3,<5.0.0)",
    "torch (>=2.6.0,<3.0.0)"
]


[tool.poetry.scripts]
article-processor = "article_processor.main:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
